dashboard: do datetime formatting in presenter
