+++ arbeitszeit/datetime_service.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
 from datetime import date, datetime
+from typing import Optional, Union
 
 
 class DatetimeService(ABC):
@@ -9,4 +10,13 @@ class DatetimeService(ABC):
 
     @abstractmethod
     def now(self) -> datetime:
+        pass
+
+    @abstractmethod
+    def format_datetime(
+        self,
+        date: Union[str, datetime],
+        zone: Optional[str] = ...,
+        fmt: Optional[str] = ...,
+    ) -> str:
         pass
