+++ arbeitszeit_flask/datetime.py
@@ -1,4 +1,7 @@
 from datetime import date, datetime
+from typing import Optional, Union
+
+from dateutil import parser, tz
 
 from arbeitszeit.datetime_service import DatetimeService
 
@@ -9,3 +12,17 @@ class RealtimeDatetimeService(DatetimeService):
 
     def today(self) -> date:
         return datetime.today().date()
+
+    def format_datetime(
+        self,
+        date: Union[str, datetime],
+        zone: Optional[str] = None,
+        fmt: Optional[str] = None,
+    ) -> str:
+        if isinstance(date, str):
+            date = parser.parse(date)
+        if zone is not None:
+            date = date.astimezone(tz.gettz(zone))
+        if fmt is None:
+            fmt = "%d.%m.%Y %H:%M"
+        return date.strftime(fmt)
