+++ arbeitszeit_web/datetime_formatter.py
@@ -1,0 +1,12 @@
+from datetime import datetime
+from typing import Optional, Protocol, Union
+
+
+class DatetimeFormatter(Protocol):
+    def format_datetime(
+        self,
+        date: Union[str, datetime],
+        zone: Optional[str] = ...,
+        fmt: Optional[str] = ...,
+    ) -> str:
+        ...
